chore: update dependencies and refactor EVM transaction handling
- Updated various dependencies in go.mod, including cosmossdk.io/log, github.com/cosmos/cosmos-db, and others to their latest versions.
- Refactored EVM transaction handling to utilize the new ethTx structure, improving type safety and consistency across the EVM module.
- Adjusted test cases to align with the new transaction structure and ensure proper validation of Ethereum transactions.

+++ cmd/gurud/config/config.go
@@ -15,32 +15,32 @@ var ChainsCoinInfo = map[uint64]evmtypes.EvmCoinInfo{
 		Denom:         ExampleChainDenom,
 		ExtendedDenom: ExampleChainDenom,
 		DisplayDenom:  ExampleDisplayDenom,
-		Decimals:      evmtypes.EighteenDecimals,
+		Decimals:      uint32(evmtypes.EighteenDecimals),
 	},
 	CosmosChainID: {
 		Denom:         "agxn",
 		ExtendedDenom: "agxn",
 		DisplayDenom:  "gxn",
-		Decimals:      evmtypes.EighteenDecimals,
+		Decimals:      uint32(evmtypes.EighteenDecimals),
 	},
 	MainnetChainID: {
 		Denom:         "agxn",
 		ExtendedDenom: "agxn",
 		DisplayDenom:  "gxn",
-		Decimals:      evmtypes.EighteenDecimals,
+		Decimals:      uint32(evmtypes.EighteenDecimals),
 	},
 	TestnetChainID: {
 		Denom:         "agxn",
 		ExtendedDenom: "agxn",
 		DisplayDenom:  "gxn",
-		Decimals:      evmtypes.EighteenDecimals,
+		Decimals:      uint32(evmtypes.EighteenDecimals),
 	},
 	// SixDecimalsChainID provides a chain ID which is being set up with 6 decimals
 	SixDecimalsChainID: {
 		Denom:         "utest",
 		ExtendedDenom: "atest",
 		DisplayDenom:  "test",
-		Decimals:      evmtypes.SixDecimals,
+		Decimals:      uint32(evmtypes.SixDecimals),
 	},
 }
